Rename known operating system family testers
