--- subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/CommonOperatingSystemFamilyTester.java
@@ -1,12 +1,0 @@
-package dev.nokee.runtime.nativebase;
-
-public interface CommonOperatingSystemFamilyTester extends
-	OperatingSystemFamilyWindowsTester,
-	OperatingSystemFamilyLinuxTester,
-	OperatingSystemFamilyMacOSTester,
-	OperatingSystemFamilyFreeBSDTester,
-	OperatingSystemFamilyIosTester,
-	OperatingSystemFamilyHPUXTester,
-	OperatingSystemFamilySolarisTester,
-	OperatingSystemFamilyUnknownTester
-{}
