+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownOperatingSystemFamilyIosTester.java
@@ -7,21 +7,21 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonOperatingSystemFamilyTester */
-interface OperatingSystemFamilySolarisTester {
+/** @see KnownOperatingSystemFamilyTester */
+interface KnownOperatingSystemFamilyIosTester {
 	OperatingSystemFamily createSubject(String name);
 
 	@ParameterizedTest(name = "can detect OS family type [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.OperatingSystemFamilyTestUtils#commonSolarisNames")
-	default void solarisFamilyNameDetected(String name) {
+	@MethodSource("dev.nokee.runtime.nativebase.OperatingSystemFamilyTestUtils#commonIosNames")
+	default void iOSFamilyNameDetected(String name) {
 		assertAll(
 			() -> assertThat("windows family", createSubject(name).isWindows(), is(false)),
 			() -> assertThat("freeBSD family", createSubject(name).isFreeBSD(), is(false)),
 			() -> assertThat("linux family", createSubject(name).isLinux(), is(false)),
 			() -> assertThat("macOS family", createSubject(name).isMacOs(), is(false)),
-			() -> assertThat("iOS family", createSubject(name).isIos(), is(false)),
+			() -> assertThat("iOS family", createSubject(name).isIos(), is(true)),
 			() -> assertThat("HP-UX family", createSubject(name).isHewlettPackardUnix(), is(false)),
-			() -> assertThat("solaris family", createSubject(name).isSolaris(), is(true))
+			() -> assertThat("solaris family", createSubject(name).isSolaris(), is(false))
 		);
 	}
 }
