+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownOperatingSystemFamilyLinuxTester.java
@@ -7,18 +7,18 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonOperatingSystemFamilyTester */
-interface OperatingSystemFamilyMacOSTester {
+/** @see KnownOperatingSystemFamilyTester */
+interface KnownOperatingSystemFamilyLinuxTester {
 	OperatingSystemFamily createSubject(String name);
 
 	@ParameterizedTest(name = "can detect OS family type [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.OperatingSystemFamilyTestUtils#commonMacOSNames")
-	default void macOSFamilyNameDetected(String name) {
+	@MethodSource("dev.nokee.runtime.nativebase.OperatingSystemFamilyTestUtils#commonLinuxNames")
+	default void linuxFamilyNameDetected(String name) {
 		assertAll(
 			() -> assertThat("windows family", createSubject(name).isWindows(), is(false)),
 			() -> assertThat("freeBSD family", createSubject(name).isFreeBSD(), is(false)),
-			() -> assertThat("linux family", createSubject(name).isLinux(), is(false)),
-			() -> assertThat("macOS family", createSubject(name).isMacOs(), is(true)),
+			() -> assertThat("linux family", createSubject(name).isLinux(), is(true)),
+			() -> assertThat("macOS family", createSubject(name).isMacOs(), is(false)),
 			() -> assertThat("iOS family", createSubject(name).isIos(), is(false)),
 			() -> assertThat("HP-UX family", createSubject(name).isHewlettPackardUnix(), is(false)),
 			() -> assertThat("solaris family", createSubject(name).isSolaris(), is(false))
