+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/KnownOperatingSystemFamilySolarisTester.java
@@ -1,6 +1,5 @@
 package dev.nokee.runtime.nativebase;
 
-import org.hamcrest.Matchers;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -8,21 +7,21 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonOperatingSystemFamilyTester */
-interface OperatingSystemFamilyWindowsTester {
+/** @see KnownOperatingSystemFamilyTester */
+interface KnownOperatingSystemFamilySolarisTester {
 	OperatingSystemFamily createSubject(String name);
 
 	@ParameterizedTest(name = "can detect OS family type [{arguments}]")
-	@MethodSource("dev.nokee.runtime.nativebase.OperatingSystemFamilyTestUtils#commonWindowsNames")
-	default void windowsFamilyNameDetected(String name) {
+	@MethodSource("dev.nokee.runtime.nativebase.OperatingSystemFamilyTestUtils#commonSolarisNames")
+	default void solarisFamilyNameDetected(String name) {
 		assertAll(
-			() -> assertThat("windows family", createSubject(name).isWindows(), is(true)),
+			() -> assertThat("windows family", createSubject(name).isWindows(), is(false)),
 			() -> assertThat("freeBSD family", createSubject(name).isFreeBSD(), is(false)),
 			() -> assertThat("linux family", createSubject(name).isLinux(), is(false)),
 			() -> assertThat("macOS family", createSubject(name).isMacOs(), is(false)),
 			() -> assertThat("iOS family", createSubject(name).isIos(), is(false)),
 			() -> assertThat("HP-UX family", createSubject(name).isHewlettPackardUnix(), is(false)),
-			() -> assertThat("solaris family", createSubject(name).isSolaris(), is(false))
+			() -> assertThat("solaris family", createSubject(name).isSolaris(), is(true))
 		);
 	}
 }
