+++ subprojects/runtime-native/src/test/groovy/dev/nokee/runtime/nativebase/UnknownOperatingSystemFamilyTester.java
@@ -7,8 +7,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-/** @see dev.nokee.runtime.nativebase.CommonOperatingSystemFamilyTester */
-interface OperatingSystemFamilyUnknownTester {
+public interface UnknownOperatingSystemFamilyTester {
 	OperatingSystemFamily createSubject(String name);
 
 	@Test
